Rename importEachEntry to importAndEditEachEntry; Add importOnlyEachEntry
- Rename importEachEntry to importAndEditEachEntry to make its name fit
its purpose. This function was designed earlier inwhich a barcode of
the current entry is checked if it exists. If YES, it will edit the
current container with new data. If NO, it will add entry as normal.
- importOnlyEachEntry: only for chemical container (chemical_storage)
It will check the barcode of the entry to see if it exists in the
current database. If YES, it will not add that entry. If NO, it will add
the entry as normal.