Add function to reload terminal page when a user logout
This commit add a reload for the terminal page when a person logout.
This is to ensure that no info of the previous chemical is shown

Why: Baylor University discover that in Terminal mode, when a user
log in, turn on stock-keeping mode, scan container, scan location
(to change location), log out, screen then shows old info even though
it was actually updated. This commit will add a page reload and therefor
no previous info will be shown