Add error alert in Terminal mode when changing chemical location without a user login; Fix set storage for multiple containers in Terminal mode
Why: in Terminal mode, if a user does not login, but with Stock-keeping
mode on, try to scan a bottle and scan location, it would APPEAR to
change location even though the change does not go into effect.

This commit adds error alert user to login first.
This commit also fixes set storage for multiple containers in Terminal mode