Made sidenav width resizeable for user that use Bootstrap4
Why: earlier version of OE does not allow the resizing of the width of
the left side bar. While most of the search functions works well with
the default setting width, in Structure search mode, the sidenav is
quite small. This commit allows users that have already enables "Use
Bootstrap4" in "Global Settings" the ability to resize the sidenav.